#7593: Add llama stress_test file which runs decode to 2k for 100 iterations.
