+++ models/demos/t3000/llama2_70b/tests/test_llama_stress_test.py
@@ -1,0 +1,160 @@
+# SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+
+# SPDX-License-Identifier: Apache-2.0
+
+import pytest
+from loguru import logger
+import torch
+from torch import nn
+import tt_lib
+import ttnn
+
+from models.demos.t3000.llama2_70b.reference.llama.llama import Llama
+from models.demos.t3000.llama2_70b.tt.llama_model_optimized import TtLlamaModel_optimized
+from models.demos.t3000.llama2_70b.tests.test_llama_perf import (
+    load_prompts_file,
+    intialize_inputs,
+)
+from models.demos.t3000.llama2_70b.tt.model_config import (
+    get_model_config,
+)
+from models.demos.t3000.llama2_70b.tt.llama_common import (
+    get_llama_path,
+    MAX_SEQ_LEN,
+    BASE_URL,
+)
+from models.utility_functions import (
+    tt2torch_tensor,
+    enable_persistent_kernel_cache,
+    disable_persistent_kernel_cache,
+    skip_for_grayskull,
+    get_devices_for_t3000,
+)
+from tqdm import tqdm
+
+
+def prepare_next_input(tokenizer, tokens, input_text_mask, cur_pos, next_token):
+    # only replace token if prompt has already been generated
+    next_token = torch.where(input_text_mask[:, cur_pos], tokens[:, cur_pos], next_token)
+    tokens[:, cur_pos] = next_token
+
+    eos_reached = (~input_text_mask[:, cur_pos]) & (next_token == tokenizer.eos_id)
+    prev_pos = cur_pos
+
+    return tokens, eos_reached, prev_pos
+
+
+def run_test_LlamaModel_stress_test(
+    devices,
+    batch,
+    seq_len,
+    model_config,
+    n_layers,
+    n_devices,
+    generation_length,
+    emulated,
+    num_users,
+):
+    devices, ckpt_dir, tokenizer_path, cache_path = get_llama_path(devices, model_config, n_devices, emulated)
+    logger.info(f"Running num_layer: {n_layers}")
+
+    generator = Llama.build(
+        ckpt_dir,
+        tokenizer_path,
+        max_seq_len=MAX_SEQ_LEN,
+        max_batch_size=num_users,
+        n_layers=1,
+        skip_model_load=False,
+    )
+    hugging_face_reference_model, tokenizer = generator.model.eval(), generator.tokenizer
+    state_dict = hugging_face_reference_model.state_dict()
+    configuration = hugging_face_reference_model.params
+
+    # Prepare input -----------------------------------------------------------------------
+    torch.manual_seed(0)
+    total_len = min(MAX_SEQ_LEN, generation_length + 1)
+    prefill_ids, ground_truth_texts = load_prompts_file(
+        tokenizer, prefill_length=32 if generation_length > 32 else 20, generation_length=generation_length
+    )
+
+    # Set up model -----------------------------------------------------------------------
+    logger.info("Moving weights to devices; might take some time...")
+    tt_model = TtLlamaModel_optimized(
+        devices,
+        state_dict,
+        BASE_URL,
+        n_layers,
+        model_config,
+        configuration,
+        batch,
+        emulated=emulated,
+        cache_path=cache_path,
+    )
+    for device in devices:
+        tt_lib.device.Synchronize(device)
+
+    del state_dict
+
+    logger.info("Starting stress test...")
+    # enable_persistent_kernel_cache()
+    for stress_test_iteration in tqdm(range(100), desc="Stress Test Progress"):
+        tokens, input_text_mask = intialize_inputs(tokenizer, prefill_ids, num_users, total_len)
+
+        start_pos = 0
+        prev_pos = start_pos
+        for cur_pos in range(start_pos + 1, total_len):
+            tt_inp_emb, prev_pos, rot_mat, attn_mask = tt_model.prepare_inputs(tokens[:, prev_pos:cur_pos], prev_pos)
+            tt_logits = tt_model(tt_inp_emb, rot_mat, prev_pos, attn_mask)
+            del tt_inp_emb, rot_mat, attn_mask
+
+            for device in devices:
+                tt_lib.device.Synchronize(device)
+
+            logits = torch.cat([tt2torch_tensor(tt_o).squeeze(1) for tt_o in tt_logits], -1)
+            logits = logits[..., : configuration.vocab_size].float()
+            del tt_logits
+
+            next_token = torch.argmax(logits, dim=-1)
+            next_token = next_token.reshape(-1)
+
+            tokens, eos_reached, prev_pos = prepare_next_input(tokenizer, tokens, input_text_mask, cur_pos, next_token)
+
+    # At the end of stress testing
+    logger.info("Completed all stress test iterations.")
+
+
+@skip_for_grayskull("Requires eth connected devices to run")
+@pytest.mark.timeout(24000000)
+@pytest.mark.models_performance_bare_metal
+@pytest.mark.parametrize(
+    "generation_length",
+    (2048,),
+    ids=["long"],
+)
+def test_Llama_stress_test(
+    generation_length,
+    all_devices,
+    use_program_cache,
+    n_layers=80,
+    n_devices=8,
+    emulated=False,
+    num_users=32,
+):
+    devices = get_devices_for_t3000(all_devices, num_devices=n_devices if not emulated else 1)
+    batch, seq_len = 32, 1
+    model_config = get_model_config(model_config_str="BFLOAT16-DRAM", num_devices=n_devices, seq_len=seq_len)
+    compute_grid_size = devices[0].compute_with_storage_grid_size()
+    if compute_grid_size.x < model_config["MAX_GRID_SIZE"][0] or compute_grid_size.y < model_config["MAX_GRID_SIZE"][1]:
+        pytest.skip(f"Requires grid size of at least {model_config['MAX_GRID_SIZE']} to run")
+
+    run_test_LlamaModel_stress_test(
+        devices,
+        batch,
+        seq_len,
+        model_config,
+        n_layers,
+        n_devices,
+        generation_length,
+        emulated,
+        num_users,
+    )
